# openapi.yaml
openapi: 3.0.3
info:
  title: x402 Insurance API
  version: 2.3.0
  description: |
    Zero-Knowledge Proof verified insurance against x402 merchant fraud.

    ## Overview
    This API provides micropayment insurance for x402 protocol users. Agents can purchase
    insurance policies to protect against merchant failures, and submit claims with zkp proofs.

    ## What's New in v2.3.0
    - **x402 V2 Payment Flow**: Full facilitator-based verification and settlement
    - **Server-Side Re-fetch Fraud Detection**: `merchant_url` required — server re-fetches to verify claims
    - **Jolt Atlas 3-Argument Prover**: `coverage_amount_units` passed to prover for payout calculation
    - **Claim Response Enrichment**: `server_verified`, `server_http_status`, `merchant_url` in responses
    - **Async Proof Generation**: Use `?async=true` on /claim to avoid timeout risk
    - **Policy Renewal**: Extend policies before expiration with /renew endpoint
    - **Agent Memory Solution**: Lookup policies by wallet address with /policies endpoint

    ## x402 Payment Protocol
    Certain endpoints require x402 payment. The payment flow:
    1. Agent makes request to protected endpoint
    2. Server responds with 402 Payment Required + payment details
    3. Agent signs payment with wallet and resubmits
    4. Server verifies payment and processes request

    ## Pricing
    - Premium: 1% of requested coverage amount
    - Coverage Range: 0.001 - 0.1 USDC (micropayments)
...

...

  contact:
...
servers:
...
paths:
...
  /claim:
    post:
      summary: Submit Fraud Claim
      description: |
        Submit a fraud claim against a merchant failure. The service will:
        1. Verify the policy is active
        2. Generate zero-knowledge proof of fraud
        3. Verify the proof
        4. Issue USDC refund to policy holder

        **Fraud Detection:**
        Fraud is detected when HTTP response shows:
        - Status code: 503, 500, 404, or other error codes
        - Empty response body
        - Missing critical headers

        **Server-Side Re-fetch (v2.3.0):**
        The server independently re-fetches `merchant_url` to verify the claimed failure.
        The `server_verified` and `server_http_status` fields in the response indicate
        whether the server confirmed the failure.

        **Jolt Prover (v2.3.0):**
        The prover now accepts 3 arguments: `http_status`, `body_length`, `coverage_amount_units`.
        The `coverage_amount_units` is used to compute the payout amount in the proof's public inputs.

        **Async Mode (NEW in v2.2.0):**
        Add `?async=true` query parameter to enable async processing:
        - Returns immediately (202 Accepted) with claim_id
        - Poll GET /claims/{claim_id} for status updates
        - Eliminates timeout risk during proof generation

      tags: [claims]
      parameters:
        - name: async
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: Enable async processing (returns 202 immediately)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ClaimRequest'
            examples:
              service_unavailable:
                summary: Merchant returned 503
                value:
                  policy_id: "123e4567-e89b-12d3-a456-426614174000"
                  http_response:
                    status: 503
                    body: ""
                    headers:
                      server: "nginx"
      responses:
        '201':
          description: Claim approved and refund issued (sync mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimResponse'
        '202':
          description: Claim accepted for processing (async mode)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AsyncClaimResponse'
        '400':
          description: Invalid claim or no fraud detected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Policy not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Proof generation or refund failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /claims/{claim_id}:
    get:
      summary: Get Claim Status
      description: |
        Retrieve the current status of a claim. Used for polling async claims.

        **Claim Statuses:**
        - `processing`: Proof generation in progress
        - `paid`: Claim approved, refund issued
        - `failed`: Claim processing failed

      tags: [claims]
      parameters:
        - name: claim_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Claim ID (UUID)
      responses:
        '200':
          description: Claim status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ClaimStatusResponse'
        '404':
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /proofs/{claim_id}:
    get:
      summary: Get Proof Data
      description: |
        Public endpoint to retrieve complete proof data for a claim.
        Returns the proof, public inputs, and verification status.

      tags: [proofs]
      parameters:
        - name: claim_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Claim ID (UUID)
      responses:
        '200':
          description: Proof data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProofData'
        '404':
          description: Claim not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
...
    ClaimRequest:
      type: object
      required:
        - policy_id
        - http_response
        - merchant_url
      properties:
        merchant_url:
          type: string
          format: uri
          description: Merchant API URL (must match policy). Used for server-side re-fetch fraud detection.
          example: "https://api.example.com/data"
        policy_id:
          type: string
          format: uuid
        http_response:
          type: object
          required:
            - status
            - body
          properties:
            status:
              type: integer
              description: HTTP status code
              example: 503
            body:
              type: string
              description: Response body (empty for fraud)
              example: ""
            headers:
              type: object
              additionalProperties:
                type: string
              description: HTTP headers
              example:
                server: "nginx"
    ClaimResponse:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
        policy_id:
          type: string
          format: uuid
        proof:
          type: string
          description: Zero-knowledge proof (hex)
          pattern: '^0x[a-fA-F0-9]+$'
        public_inputs:
          type: array
          items:
            type: integer
          description: |
            Public inputs to the proof:
            [0]: is_fraud (1 = yes, 0 = no)
            [1]: http_status
            [2]: body_length
            [3]: suggested_payout
        payout_amount:
          type: number
          description: Actual payout in USDC
        refund_tx_hash:
          type: string
          pattern: '^0x[a-fA-F0-9]{64}$'
          description: Blockchain transaction hash
        status:
          type: string
          enum: [paid, pending, rejected]
        server_verified:
          type: boolean
          description: Whether server independently re-fetched the merchant URL to verify failure
        server_http_status:
          type: integer
          description: HTTP status code from server's independent re-fetch of merchant URL
        merchant_url:
          type: string
          format: uri
          description: The merchant API URL that was checked
        proof_url:
          type: string
          description: URL to retrieve full proof data
    AsyncClaimResponse:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
          description: Claim ID for polling status
        status:
          type: string
          enum: [processing]
          description: Initial status (always "processing")
        poll_url:
          type: string
          description: URL to poll for claim status
          example: "/claims/123e4567-e89b-12d3-a456-426614174000"
        estimated_completion_seconds:
          type: integer
          description: Estimated time to completion
          example: 20
    ClaimStatusResponse:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [processing, paid, failed]
          description: Current claim status
        payout_amount:
          type: number
          nullable: true
          description: Payout amount (null if still processing)
        refund_tx_hash:
          type: string
          nullable: true
          pattern: '^0x[a-fA-F0-9]{64}$'
          description: Transaction hash (null if still processing)
        error:
          type: string
          nullable: true
          description: Error message (only if status is "failed")
        proof:
          type: string
          nullable: true
          description: Zero-knowledge proof (available when paid)
        public_inputs:
          type: array
          nullable: true
          items:
            type: integer
          description: Public inputs (available when paid)
...
    ProofData:
      type: object
      properties:
        claim_id:
          type: string
          format: uuid
        proof:
          type: string
        public_inputs:
          type: array
          items:
            type: integer
        http_status:
          type: integer
        http_body_hash:
          type: string
        http_headers:
          type: object
        verification_result:
          type: boolean
        payout_amount:
          type: number
        refund_tx_hash:
          type: string
        recipient_address:
          type: string
        status:
          type: string
        created_at:
          type: string
          format: date-time
        paid_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message